+++ test/test_ecc_trivariate_tricycle.jl
@@ -77,6 +77,7 @@
             @test code_k(c) == k == code_k(stab)
             @test stab_looks_good(stab, remove_redundant_rows=true) == true
             @test iszero(mod.(metacheck_matrix_z(c)*parity_matrix_z(c), 2))
+            Hx = parity_matrix_x(c)
             Hz = parity_matrix_z(c)
             Mz = metacheck_matrix_z(c)
             @test Hz != Mz
@@ -92,6 +93,7 @@
             @test code_k(c) == k == code_k(stab)
             @test stab_looks_good(stab, remove_redundant_rows=true) == true
             @test iszero(mod.(metacheck_matrix_z(c)*parity_matrix_z(c), 2))
+            Hx = parity_matrix_x(c)
             Hz = parity_matrix_z(c)
             Mz = metacheck_matrix_z(c)
             @test Hz != Mz